refactor(extensions)!: move YfmHtmlBlock extension from yfm to additional folder
