+++ demo/Playground.tsx
@@ -23,8 +23,8 @@ import {Extension} from '../src/cm/state';
 import {FoldingHeading} from '../src/extensions/additional/FoldingHeading';
 import {Math} from '../src/extensions/additional/Math';
 import {Mermaid} from '../src/extensions/additional/Mermaid';
-import {YfmHtmlBlock} from '../src/extensions/yfm/YfmHtmlBlock';
-import {getSanitizeYfmHtmlBlock} from '../src/extensions/yfm/YfmHtmlBlock/utils';
+import {YfmHtmlBlock} from '../src/extensions/additional/YfmHtmlBlock';
+import {getSanitizeYfmHtmlBlock} from '../src/extensions/additional/YfmHtmlBlock/utils';
 import {cloneDeep} from '../src/lodash';
 import {CodeEditor} from '../src/markup/editor';
 import type {FileUploadHandler} from '../src/utils/upload';
