+++ src/extensions/additional/YfmHtmlBlock/YfmHtmlBlock.test.ts
@@ -2,7 +2,7 @@ import {builders} from 'prosemirror-test-builder';
 
 import {createMarkupChecker} from '../../../../tests/sameMarkup';
 import {ExtensionsManager} from '../../../core';
-import {BaseNode, BaseSchemaSpecs} from '../../../extensions/specs';
+import {BaseNode, BaseSchemaSpecs} from '../../specs';
 
 import {YfmHtmlBlockSpecs} from './YfmHtmlBlockSpecs';
 import {YfmHtmlBlockAttrs, yfmHtmlBlockNodeName} from './const';
